+++ CHANGELOG.md
@@ -4,6 +4,9 @@ conda-docker Change Log
 
 
 <!-- current developments -->
+**Added:**
+
+* Adding support for setting docker v1 layer config e.g. ENV, LABEL, CMD, ENTRYPOINT etc.
 
 ## v0.0.1
 **Added:**